web discovery and path assertion refactoring
web discovery was refactered to now only report back by returning a graph.
the ensures that it can be used seperately from the path assertion and from travharv itself.
did some refactoring also in the path assertion to make it more readable and easier to understand.
Also fixed the fact that the path assertion was not working correctly when the path was not found.
the approach we have now is 1shot , if no path is found for a given url, we will not try to find it again for another found subjects in the same url.
The execution reports have also been modified to prepare better for putting the provenance information in the report.
The tests have been updated to reflect the changes in the code.